refactor: MMS-43 pull changes from dev, write logic without static in domain
